Updated tests with separate data headers to use best practice and separate out the json link code
