+++ tests/src/apache_builds_cpp_comments_test.cpp
@@ -12,7 +12,7 @@
 
 #include "defines.h"
 
-#include "apache_builds.h"
+#include "apache_builds_json.h"
 #include "daw/json/daw_json_link.h"
 
 #include <daw/cpp_17.h>
