+++ tests/src/nativejson_bench_basic.cpp
@@ -8,9 +8,9 @@
 
 #include "defines.h"
 
-#include "citm_test.h"
+#include "citm_test_json.h"
 #include "geojson.h"
-#include "twitter_test.h"
+#include "twitter_test_json.h"
 
 #include "daw/json/daw_json_link.h"
