+++ tests/src/twitter_test.cpp
@@ -9,7 +9,7 @@
 #include "defines.h"
 
 #include "daw/json/daw_json_link.h"
-#include "twitter_test.h"
+#include "twitter_test_json.h"
 
 #include <daw/cpp_17.h>
 #include <daw/daw_benchmark.h>
